Atlantic Records Website
Key Features:

HTML5 and CSS3 for replicating the Atlantic Records website's structure and style.
Added image slideshow feature to the homepage, providing an eye-catching visual element.
Responsive design ensures an optimal viewing experience on various devices.
Navigation menus, headers, and footers remain consistent with the original site.

Disclaimer:
This project is for educational and personal purposes only and is not affiliated with or endorsed by Atlantic Records.